fix: give mac developers feedback when `yarn` install fails because of an incorrect python dependency
On macOS, this script checks if Python 3.10 is installed and accessible to node-gyp.

I ran into a problem trying to `yarn` install, with a system Python version of `3.12.2`, but ran into the error `ModuleNotFoundError: No module named 'distutils'`. Since node-gyp relies on `distutils`, which is removed in Python `3.12`, you need to use a Python version that still includes `distutils`.